+++ src/main/lib/worldInflate.ts
@@ -1,0 +1,6 @@
+import { expose } from 'threads/worker';
+import * as pako from 'pako';
+
+expose(function (data, size) {
+	return pako.inflate(data, new Uint16Array(size));
+});